rename packages/files
